requirements.txt
pillow==10.3.0
aiohttp==3.9.4
zhipuai==2.0.1
PyQt5==5.15.10
PyQtWebEngine==5.15.6
FastAPI==0.110.1
bs4==0.0.2
uvicorn==0.29.0
websockets==12.0
httpx~=0.27.0
pyjwt~=2.8.0
beautifulsoup4~=4.12.3
lxml~=5.2.1
pyinstaller==5.13.0